Moving streaming code in the common place
